+++ source/common/vue/form/elements/Button.vue
@@ -75,5 +75,7 @@ export default defineComponent({
 </script>
 
 <style lang="less">
-
+body .form-control button {
+  margin: 5px;
+}
 </style>